+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Improved checks for "OtherController" response
+
 ## [1.24.2] - 2023-08-30
 
 ### Changed